extract the sequences to animated component
